change move AddRMQObjectToRequestMiddleware to middlewares
