+++ src/middlewares/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 from .http_proxy_middleware import HttpProxyMiddleware
 from .log_errors_middleware import LogErrorsMiddleware
+from .add_rmq_object_to_request_middleware import AddRMQObjectToRequestMiddleware
