+++ src/middlewares/add_rmq_object_to_request_middleware.py
